- Fix #4236: Crash in multi line string - Remove dead code - Remove file system completions in string literals
