+++ Python/Tests/Core/EditorTests.cs
@@ -518,66 +518,6 @@ In [3]: if True:
 
         #endregion
 
-        private static StringLiteralCompletionList.EntryInfo MakeEntryInfo(string rootpath, string filename, string insertionText = null, string fullpath = null, bool? isFile = null) {
-            var realIsFile = isFile ?? !string.IsNullOrEmpty(Path.GetExtension(filename));
-            return new StringLiteralCompletionList.EntryInfo {
-                Tooltip = fullpath ?? Path.Combine(rootpath, filename),
-                InsertionText = insertionText ?? (Path.Combine(rootpath, filename) + (realIsFile ? "" : "\\")),
-                Caption = filename,
-                IsFile = realIsFile
-            };
-        }
-
-        [TestMethod, Priority(0)]
-        public void StringCompletionFileEntries() {
-            var cwd = TestData.GetPath("TestData");
-            var user = TestData.GetPath("TestData\\Databases");
-
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo(cwd + "\\AbsolutePaths", cwd, user),
-                MakeEntryInfo(cwd, "AbsolutePath"),
-                MakeEntryInfo(cwd, "AbsolutePath.sln"),
-                MakeEntryInfo(cwd, "HelloWorld"),
-                MakeEntryInfo(cwd, "HelloWorld.sln")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo(cwd + "\\AbsolutePath\\", cwd, user),
-                MakeEntryInfo(cwd + "\\AbsolutePath", "AbsolutePath.pyproj")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo("./AbsolutePaths", cwd, user),
-                MakeEntryInfo(cwd, "AbsolutePath", ".\\AbsolutePath\\"),
-                MakeEntryInfo(cwd, "AbsolutePath.sln", ".\\AbsolutePath.sln"),
-                MakeEntryInfo(cwd, "HelloWorld", ".\\HelloWorld\\"),
-                MakeEntryInfo(cwd, "HelloWorld.sln", ".\\HelloWorld.sln")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo(".\\Ab", cwd, user),
-                MakeEntryInfo(cwd, "AbsolutePath", ".\\AbsolutePath\\"),
-                MakeEntryInfo(cwd, "AbsolutePath.sln", ".\\AbsolutePath.sln"),
-                MakeEntryInfo(cwd, "HelloWorld", ".\\HelloWorld\\"),
-                MakeEntryInfo(cwd, "HelloWorld.sln", ".\\HelloWorld.sln")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo("~/Ab", cwd, user),
-                MakeEntryInfo(user, "V27", "~\\V27\\"),
-                MakeEntryInfo(user, "Readme.txt", "~\\Readme.txt")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo("~\\Ab", cwd, user),
-                MakeEntryInfo(user, "V27", "~\\V27\\"),
-                MakeEntryInfo(user, "Readme.txt", "~\\Readme.txt")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo("~\\V27\\", cwd, user),
-                MakeEntryInfo(user + "\\V27", "ntpath.idb", "~\\V27\\ntpath.idb"),
-                MakeEntryInfo(user + "\\V27", "os.idb", "~\\V27\\os.idb")
-            );
-            AssertUtil.ContainsAtLeast(
-                StringLiteralCompletionList.GetEntryInfo("Ab", cwd, user)
-            );
-        }
-
         #endregion Test Cases
 
         #region Helpers
